feat: create a seperate retriever to fetch conflicting rules from other node_set if any. initial commit.
