+++ cognee/modules/search/types/SearchType.py
@@ -7,6 +7,7 @@ class SearchType(Enum):
     CHUNKS = "CHUNKS"
     RAG_COMPLETION = "RAG_COMPLETION"
     GRAPH_COMPLETION = "GRAPH_COMPLETION"
+    GRAPH_COMPLETION_WITH_CONFLICTS = "GRAPH_COMPLETION_WITH_CONFLICTS"
     GRAPH_SUMMARY_COMPLETION = "GRAPH_SUMMARY_COMPLETION"
     CODE = "CODE"
     CYPHER = "CYPHER"
